refactor(agent-worker): split shell-command parsing out of tool-policy
tool-policy.ts mixed two unrelated concerns: shell command lexing (quote
blanking, sub-command splitting, argument-boundary matching) and tool
policy evaluation. It is the most frequently opened file in the repo, so
the cost of reading past one concern to reach the other was paid often.

Moves the lexer to bash-command-parser.ts. All moved symbols were private,
so the exported surface is unchanged (six exports, identical to before).

667 -> 347 + 326 lines.

(cherry picked from commit f29417a7239c522f84416f87a931879c1a4021e3)
